round frp

+++ eumetsat-inflow/eumetsat.py
@@ -174,7 +174,7 @@ def getNearestFire(consumer_key, consumer_secret, ref_lat, ref_lon, bbox):
 				"distance": round(distance, 1),
 				"latitude": float(latitude),
 				"longitud": float(longitud),
-				"frp": float(frp)
+				"frp": int(frp)
 		}
 	except Exception as e:
 		logging.error(f"Error at getNearestFire: {str(e)}")
